requirements.txt
streamlit
pandas
pytesseract
pillow
PyMuPDF
openpyxl
